+++ release.nix
@@ -1,7 +1,8 @@
-{ reflex-platform ? import ./reflex-platform
+{ reflex-platform ? import ./dep/reflex-platform
 }:
 let
-  pkgs = (reflex-platform {}).nixpkgs;
+  rp = reflex-platform {};
+  pkgs = rp.nixpkgs;
   supportedSystems = [ "x86_64-linux" "x86_64-darwin" ];
   inherit (pkgs) lib;
   haskellLib = pkgs.haskell.lib;
@@ -11,13 +12,14 @@ let
       ver = "5.38";
       sha256 = "0kcd3ln9xmc62ka0i7habzvjjar8z63mlvl15rdhf8hqmda0b7r7";
     } {};
+    reflex = self.callCabal2nix "reflex" (rp.hackGet ./dep/reflex) {};
   };
   ghcs = lib.genAttrs supportedSystems (system: let
     rp = reflex-platform { inherit system; __useNewerCompiler = true; };
     rpGhc = rp.ghc.override {
       overrides = commonOverrides;
     };
-    nixGhc945 = (import ./nixpkgs { inherit system; }).haskell.packages.ghc945.override {
+    nixGhc945 = (import ./dep/nixpkgs { inherit system; }).haskell.packages.ghc945.override {
       overrides = self: super: commonOverrides self super // {
         hlint = self.callHackageDirect {
           pkg = "hlint";
@@ -40,7 +42,7 @@ let
         } {};
       };
     };
-    nixGhc961 = (import ./nixpkgs { inherit system; }).haskell.packages.ghc961.override {
+    nixGhc961 = (import ./dep/nixpkgs { inherit system; }).haskell.packages.ghc961.override {
       overrides = self: super: {
         patch = self.callHackageDirect {
           pkg = "patch";